refactor: typting 로직 분리
